+++ website/api/controllers/webhooks/receive-from-github.js
@@ -109,9 +109,9 @@ module.exports = {
       'escomeau',
       'cdcme',
       'kevinmalkin12',
-      'Illbjorn',
       'karmine05',
       'ericswenson0',
+      'kitzy',
     ];
 
     let GREEN_LABEL_COLOR = 'C2E0C6';// « Used in multiple places below.  (FUTURE: Use the "+" prefix for this instead of color.  2022-05-05)